records.py module
Changed the import position otherwise only skyrim loads (oblivion module not
seen)

--------------------- Mopy/bash/game/oblivion/__init__.py ---------------------
index b11775d..d8f7cf0 100644
@@ -25,7 +25,6 @@
 """This modules defines static data for use by bush, when
    TES IV: Oblivion is set at the active game."""

-from game.oblivion.records import RecordHeader
 from oblivion_const import bethDataFiles, allBethFiles
 from ... import brec
 from ... import bush
@@ -1678,6 +1677,7 @@ class esp:

     recordTypes = set(topTypes + 'GRUP,TES4,ROAD,REFR,ACHR,ACRE,PGRD,LAND,INFO'.split(','))

+from records import * # MUST BE HERE otherwise all hell breaks loose
 #------------------------------------------------------------------------------
 # Record Elements    ----------------------------------------------------------
 #------------------------------------------------------------------------------
