+++ Mopy/bash/game/oblivion/__init__.py
@@ -25,10 +25,8 @@
 """This modules defines static data for use by bush, when
    TES IV: Oblivion is set at the active game."""
 
-import struct
 from oblivion_const import bethDataFiles, allBethFiles
 from ... import brec
-from ... import bolt
 from ... import bush
 from ...brec import *
 from ...bolt import Flags, DataDict, StateError
@@ -1679,56 +1677,7 @@ class esp:
 
     recordTypes = set(topTypes + 'GRUP,TES4,ROAD,REFR,ACHR,ACRE,PGRD,LAND,INFO'.split(','))
 
-#--Mod I/O
-class RecordHeader(brec.BaseRecordHeader):
-    size = 20
-
-    def __init__(self,recType='TES4',size=0,arg1=0,arg2=0,arg3=0,*extra):
-        self.recType = recType
-        self.size = size
-        if recType == 'GRUP':
-            self.label = arg1
-            self.groupType = arg2
-            self.stamp = arg3
-        else:
-            self.flags1 = arg1
-            self.fid = arg2
-            self.flags2 = arg2
-        self.extra = extra
-
-    @staticmethod
-    def unpack(ins):
-        """Returns a RecordHeader object by reading the input stream."""
-        type,size,uint0,uint1,uint2 = ins.unpack('=4s4I',20,'REC_HEADER')
-        #--Bad?
-        if type not in esp.recordTypes:
-            raise brec.ModError(ins.inName,u'Bad header type: '+repr(type))
-        #--Record
-        if type != 'GRUP':
-            pass
-        #--Top Group
-        elif uint1 == 0: # groupType == 0 (Top Group)
-            str0 = struct.pack('I',uint0)
-            if str0 in esp.topTypes:
-                uint0 = str0
-            elif str0 in esp.topIgTypes:
-                uint0 = esp.topIgTypes[str0]
-            else:
-                raise brec.ModError(ins.inName,u'Bad Top GRUP type: '+repr(str0))
-        return RecordHeader(type,size,uint0,uint1,uint2)
-
-    def pack(self):
-        """Returns the record header packed into a string for writing to file."""
-        if self.recType == 'GRUP':
-            if isinstance(self.label,str):
-                return struct.pack('=4sI4sII',self.recType,self.size,self.label,self.groupType,self.stamp)
-            elif isinstance(self.label,tuple):
-                return struct.pack('=4sIhhII',self.recType,self.size,self.label[0],self.label[1],self.groupType,self.stamp)
-            else:
-                return struct.pack('=4s4I',self.recType,self.size,self.label,self.groupType,self.stamp)
-        else:
-            return struct.pack('=4s4I',self.recType,self.size,self.flags1,self.fid,self.flags2)
-
+from records import * # MUST BE HERE otherwise all hell breaks loose
 #------------------------------------------------------------------------------
 # Record Elements    ----------------------------------------------------------
 #------------------------------------------------------------------------------
